paddlepaddle/requirements.txt
albucore==0.0.13
albumentations==1.4.10
annotated-types==0.7.0
anyio==4.6.2.post1
astor==0.8.1
babel==2.16.0
bce-python-sdk==0.9.23
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cssselect==1.2.0
cssutils==2.11.1
cycler==0.12.1
cython==3.0.11
decorator==5.1.1
easydict==1.13
et-xmlfile==2.0.0
exceptiongroup==1.2.2
faiss-cpu==1.8.0.post1
fire==0.7.0
flask==3.1.0
flask-babel==4.0.0
fonttools==4.55.0
future==1.0.0
gast==0.3.3
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
imageio==2.36.0
imgaug==0.4.0
itsdangerous==2.2.0
jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy-loader==0.4
lmdb==1.5.1
lxml==5.3.0
markupsafe==3.0.2
matplotlib==3.9.2
more-itertools==10.5.0
networkx==3.4.2
numpy==1.24.4
opencv-contrib-python==4.10.0.84
opencv-python==4.6.0.66
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==24.2
paddleclas==2.6.0
paddleocr==2.9.1
paddlepaddle-gpu==2.6.2
pandas==2.2.3
pdf2image==1.17.0
pillow==11.0.0
premailer==3.10.0
prettytable==3.12.0
protobuf==5.28.3
psutil==6.1.0
pyclipper==1.3.0.post6
pycryptodome==3.21.0
pydantic==2.9.2
pydantic-core==2.23.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
pyyaml==6.0.2
rapidfuzz==3.10.1
rarfile==4.2
requests==2.32.3
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.5.0
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tomli==2.1.0
tqdm==4.67.0
typing-extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
visualdl==2.5.3
wcwidth==0.2.13
werkzeug==3.1.3
